// src/app/(client)/photos/types.ts
export type Photo = {
    id: number;
    width: number;
    height: number;
    url: string;
    photographer: string;
    photographer_url: string;
    photographer_id: number;
    avg_color: string;
    src: {
        original: string;
        medium: string;
    }
    liked: boolean;
    alt: string;
}